Extract BitmapDrawable scaling/conversion code to separate classes
No functional changes.
